api/requirements.txt
Django==4.1.7
psycopg2==2.9.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
gunicorn==20.1.0
pytz==2022.7.1
whitenoise==6.3.0
certifi==2022.12.7
dj-database-url==1.2.0
dj-rest-auth==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-light==0.1.0.post3
drf-yasg==1.21.5
swagger-spec-validator==3.0.3

jupyter==1.0.0
ipython==8.10.0